% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_break_signal.R
\name{break_signal}
\alias{break_signal}
\title{Break Signal}
\usage{
break_signal(
  y = integer(0),
  k = 0,
  grp = NULL,
  obs_ctrl = list(min_size = 1L, max_k = 2L),
  best_k = numeric(0),
  alt_k = numeric(0),
  k_sz = numeric(0),
  score = numeric(0),
  data = NULL
)
}
\description{
Functions \code{break_signal()} and \code{signal_processor()} are designed to select the best break of a “signal”, observed differences in the values of an ordered sequence. A “break” is simply a change in sign of the differences: the “best” break is the one that indicates the process that generates of observations has changed state.
}
\section{Slots}{

\describe{
\item{\code{y}}{The un-transformed, monotonically ordered vector}

\item{\code{grp}}{A vector containing group assignments along \code{y}}

\item{\code{obs_ctrl}}{(list) Observation data control list having the following recognized elements:\cr
\itemize{
\item{\code{max_k}: The upper limit for breaks to allow under consideration. This should primarily be set to control the effects of extreme values that appear frequently enough to be non-trivial but are due to confounding factors not related to the generative process under study.}
\item{\code{min_size}: The minimum observation size at each k to allow}
}}

\item{\code{best_k,alt_k}}{(numeric) Entropy-based optimized (and next best) break in the signal (passively set by \code{\link{signal_processor}})}

\item{\code{k_sz}}{(numeric) The observation size at each \code{k} (passively set by \code{\link{signal_processor}})}

\item{\code{score}}{(numeric[]) The scoring vector at each \code{k} (passively set by \code{\link{signal_processor}})}

\item{\code{data}}{The generated data for which scores are generated (passively set by \code{\link{signal_processor}})}

\item{\code{plot}}{A \code{\link[plotly]{plot_ly}} visualization after break optimization (passively set by \code{\link{signal_processor}})}
}}

