% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. cross.time.R
\name{cross.time}
\alias{cross.time}
\title{Cross-Compare Temporal Boundaries}
\usage{
cross.time(s0, s1, e0, e1, control = list(-Inf, Inf), chatty = FALSE, ...)
}
\arguments{
\item{s0}{A numeric/date-coded vector containing the temporal lower boundary of the starting event duration}

\item{s1}{A numeric/date-coded vector containing the temporal upper boundary of the starting event duration}

\item{e0}{A numeric/date-coded vector containing the temporal lower boundary of the ending event duration}

\item{e1}{A numeric/date-coded vector containing the temporal upper boundary of the ending event duration}

\item{control}{A length-2 sorted list with values indicating the range of allowable values for internal variable \code{beta} (the difference between ends of 'to' events and beginnings of 'from' events)}

\item{chatty}{(logical | \code{FALSE}) Verbosity flag}

\item{...}{(Not used)}
}
\value{
A \code{\link[data.table]{data.table}} object having the following fields:
\describe{
\item{\code{mGap: }}{Metric describing the difference between the following temporal boundaries: \code{TO.start}, \code{FROM.end}}
\item{\code{mSt: }}{Metric describing the difference between the following temporal boundaries: \code{TO.start}, \code{FROM.start}}
\item{\code{mEd: }}{Metrics describing the difference between the following temporal boundaries: \code{TO.end}, \code{FROM.end}}
\item{\code{from.len: }}{The duration of time of each "from" event: units are one the scale of the smallest increment of time represented (e.g., calendar dates in days will have lengths expressed in days)}
\item{\code{to.len: }}{The duration of time of each "to" event: units are one the scale of the smallest increment of time represented (e.g., calendar dates in days will have lengths expressed in days)}
\item{\code{epsilon: }}{A complex number (e.g., \code{1337 + 0.90210i}) describing the relational changes from one event to another with interpretation based on whether or not the real and imaginary parts are \code{> 0}, \code{< 0}, or \code{ == 0}:
	\tabular{lll}{
	  Re \tab Im \tab Desc \cr
	  {> 0} \tab {0} \tab Disjoint \cr
	  {0} \tab {> 0} \tab Concurrency \cr
	  {> 0} \tab {> 0} \tab Full Concurrency \cr
	  {0} \tab {0} \tab Continuity \cr
	  { } \tab {< 1} \tab \code{to} event shorter than \code{from} event \cr
	  { } \tab {= 1} \tab \code{to} event same length as \code{from} event \cr
	  { } \tab {> 1} \tab \code{to} event longer than \code{from} event \cr
	}
}
\item{\code{epsilon.desc: }}{A plain-language description of \code{epsilon}}
}
}
\description{
Given a four-element vector of start and end coordinates of two events, \code{cross.time()} compares the distances among the upper and lower boundaries of pairs of event vectors. This includes "like" boundary comparison (e.g., start #2 - start#1) and contrary boundary comparison (e.g. start #2 - end #1).

\code{cross.time()} \code{\link[memoise]{memoise}}s arguments \code{s0}, \code{s1}, \code{e0}, and \code{e1} with a disk cache defined as \code{\link[cachem]{cache_disk}}\code{(dir = tempdir(), max_age = Inf, max_n = Inf, destroy_on_finalize = FALSE)}.  A future version will allow the cache to be customized in a post-hoc manner.
}
