% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Cross-Compare Temporal Boundaries}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{s0}{A numeric/date-coded vector containing the temporal lower boundary of the starting event duration}

\item{s1}{A numeric/date-coded vector containing the temporal upper boundary of the starting event duration}

\item{e0}{A numeric/date-coded vector containing the temporal lower boundary of the ending event duration}

\item{e1}{A numeric/date-coded vector containing the temporal upper boundary of the ending event duration}

\item{control}{A length-2 sorted list with values indicating the range of allowable values for '.beta' (the difference between ends of 'to' events and beginnings of 'from' events)}

\item{events.ascending}{(logical | TRUE) \code{FALSE} assumes events provided are in descending order}

\item{chatty}{(logical | \code{FALSE}) Use \code{chatty = TRUE} to see messages related to the execution.}

\item{nosy}{(logical | FALSE) Used for debugging purposes}
}
\value{
A \code{\link[data.table]{data.table}} object (??event.vector.space)
}
\description{
Given a four-element vector of start and end coordinates of two events, \code{cross.time()} compares the distances among the upper and lower boundaries of pairs of event vectors. This includes "like" boundary comparison (e.g., start #2 - start#1) and contrary boundary comparison (e.g. start #2 - end #1).
}
