% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_evs_exclude.blender.R
\name{evs_exclude.blender}
\alias{evs_exclude.blender}
\title{Event-Pair Exclusions}
\usage{
evs_exclude.blender(x, y)
}
\arguments{
\item{x}{See \code{\link[base]{expand.grid}}}

\item{y}{See \code{\link[base]{expand.grid}}}
}
\value{
A vector of pairs of event labels to be excluded in the call to \code{$exclude.mix()}
}
\description{
\code{evs_exclude.blender} is a wrapper for \code{\link[base]{expand.grid}} with some post-processing via \code{\link[purrr]{array_branch}} and \code{\link[purrr]{reduce}}.
}
\examples{
evs_exclude.blender(
  x = c("A", "B")
  , y = c("A", "D", "Z")
  )

}
