% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. otherFunctions.R
\name{evs_retrace}
\alias{evs_retrace}
\title{Retrace Event Source Data}
\usage{
evs_retrace(self, ...)
}
\arguments{
\item{self}{An R object of class "event.vector.space"}

\item{...}{(\code{\link[rlang]{dots_list}}) Names of event graphs found in \code{self$evt_graphs}}
}
\value{
Because of the reference semantics of R6 classes, for each name given in \code{...}, graph updates are in place: \code{self} is returned invisibly.
}
\description{
\code{evs_retrace} creates vertex attribute \code{trace} and populates it with an expression that retrieves the source record when evaluated.
}
