% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evs_src_01}
\alias{evs_src_01}
\title{Event Vector Space Test Data #1}
\format{
\subsection{\code{evs_src_01}}{

A \code{\link[data.table]{data.table}} objects prefixed as 'evs_src_' with the following structure:
\describe{
\item{join_key}{An integer key to be used for joining datasets}
\item{date.start}{A date column representing the start of an event}
\item{date.end}{A date column representing the end of an event}
\item{src}{A character column representing the source of the event}
\item{X_001, X_002, ...}{Numeric columns representing attributes of the events, with random values between 0 and 1}
}
}
}
\source{
Synthesized by the package author.
}
\usage{
evs_src_01
}
\description{
One of three test datasets that can be used to explore
package functionality.
}
\keyword{datasets}
