% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_signal_processor.R
\name{exec_algorithm}
\alias{exec_algorithm}
\title{Execute Algorithm}
\usage{
exec_algorithm(Data, obs_ctrl, ..., chatty = FALSE)
}
\arguments{
\item{Data}{The input data}

\item{obs_ctrl}{(See \code{\link{break_signal}})}

\item{...}{Arguments for internal use}

\item{chatty}{(logical) Should the function be verbose?}
}
\value{
A \code{\link[data.table]{data.table}} object containing scoring information.
}
\description{
\code{exec_algorithm} executes the algorithm for a given data set and number of cross-assignment folds.
The following functions are called in the order listed below:\cr
\enumerate{
\item{\code{\link{info_encoder}}}
\item{\code{\link{geo_pmf}}}
\item{\code{\link{score_algorithm_output}}}
}
}
\seealso{
Other Signal Processor Functions: 
\code{\link{geo_pmf}()},
\code{\link{info_encoder}()},
\code{\link{score_algorithm_output}()},
\code{\link{signal_processor}()}
}
\concept{Signal Processor Functions}
