% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_signal_processor.R
\name{geo_pmf}
\alias{geo_pmf}
\title{Geometric Probability Mass Function}
\usage{
geo_pmf(x, p)
}
\arguments{
\item{x}{The cumulative sum of differences, representing the number of attempts before success.\cr These values imply contiguous parts of the series before a break.}

\item{p}{The cumulative proportion of the current group's difference-proportion mapped from \code{response.pmf}.}
}
\value{
\eqn{p(1 - p)^x}
}
\description{
Geometric Probability Mass Function
}
\seealso{
Other Signal Processor Functions: 
\code{\link{exec_algorithm}()},
\code{\link{info_encoder}()},
\code{\link{score_algorithm_output}()},
\code{\link{signal_processor}()}
}
\concept{Signal Processor Functions}
