% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_signal_processor.R
\name{info_encoder}
\alias{info_encoder}
\title{Encode Shannon Information}
\usage{
info_encoder(i, info.only = FALSE, data.only = info.only)
}
\arguments{
\item{i}{(numeric) The input vector of differences}

\item{info.only}{(logical) \code{TRUE} returns the information column only}

\item{data.only}{(logical) \code{TRUE} returns columns \code{series} and \code{cyl} only, and \code{FALSE} returns everything}
}
\value{
If \code{info.only} is \code{TRUE}, a numeric vector of Shannon information is returned, otherwise a data frame with columns \code{series}, \code{cyl}, and \code{info} is returned.
}
\description{
\code{info_encoder} converts the \code{i} into distinct series based on monotonically occurring breaks.\cr
The cumulative proportionality within each series is converted into Shannon information as "bits".
}
\note{
\code{info_encoder()} requires the input to be the differences of an \emph{ordered} vector
}
\seealso{
Other Signal Processor Functions: 
\code{\link{exec_algorithm}()},
\code{\link{geo_pmf}()},
\code{\link{score_algorithm_output}()},
\code{\link{signal_processor}()}
}
\concept{Signal Processor Functions}
