% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_retrace.evs.R
\name{retrace.evs}
\alias{retrace.evs}
\title{Retrace Events to Source}
\usage{
retrace.evs(event_graph, evs)
}
\arguments{
\item{event_graph}{A graph contained within \code{$evt_graphs} (use the \code{`[[`} primitive if using a numeric index; otherwise, use \code{`$`})}

\item{evs}{An object of class \code{event.vectors}}
}
\value{
A list of distinct rows that map to the source event data.  Event graph vertex names can be used to subset the output.
}
\description{
\code{retrace.evs} uses the input graph and reference \code{event.vectors} object to return the source data rows corresponding to the event vertices.
}
