% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_signal_processor.R
\name{score_algorithm_output}
\alias{score_algorithm_output}
\title{Score Algorithm Output}
\usage{
score_algorithm_output(algo_output, obs_ctrl)
}
\arguments{
\item{algo_output}{(data frame) The output of \code{\link{exec_algorithm}}.}

\item{obs_ctrl}{(list) (See \code{\link{break_signal}})}
}
\value{
A \code{\link[data.table]{data.table}} with the following columns:\cr
\describe{
\item{k}{The vector of \code{k} values}
\item{d_Idev}{First-order difference of \code{Idev}}
\item{d2_Idev}{Second-order difference of \code{Idev}}
\item{k_score}{The break score -> information deviation equation/model}
\item{k_sz}{The number of observations in each \code{k}}
\item{geo_pmf.max}{The maximal value of the geometric probability mass function returned by \code{geo_pmf}}
\item{Idev_wmean}{The weighted mean of \code{Idev} using as weights the geometric PMF (\code{geo_pmf}) in order to find the optimal \code{k}}
\item{d_kscore, d_Idev_wmean}{For each \code{k}, the slope of \code{k_score} and \code{Idev_wmean}, respectively}
\item{tot_score}{The mutual relative proportionality of curvatures over \code{k} and \code{Idev_wmean}. The relative proportionality of \code{d2_Idev_wmean} is subtracted from 1 to invert the proportionality.  This makes \code{tot_score} the product of the \emph{minimum} of one curvature and the \emph{maximum} of the othe (gradient descent is not used to determine the optimal \code{k} because \code{k} is a discrete value.)}
\item{best_k}{The \code{k} value with the highest \code{tot_score}}
\item{alt_k}{The \code{k} value that is not the \code{best_k} having a cumulative proportion of \code{tot_score} >= 0.9}
}
}
\description{
\code{score_algorithm_output} is a scoring function definition used in function \code{\link{exec_algorithm}}.
}
\seealso{
Other Signal Processor Functions: 
\code{\link{exec_algorithm}()},
\code{\link{geo_pmf}()},
\code{\link{info_encoder}()},
\code{\link{signal_processor}()}
}
\concept{Signal Processor Functions}
