% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_continuity.R
\name{continuity}
\alias{continuity}
\title{Continuity Creator}
\usage{
continuity(
  data,
  map_fields,
  time_fields,
  timeout = 0,
  boundary_name = "window",
  archipelago = TRUE,
  show.all = FALSE
)
}
\arguments{
\item{data}{(object): The source dataset, including all non-sessioning fields desired}

\item{map_fields}{(string): A vector of strings or symbols indicating the field names that will partition \code{data}}

\item{time_fields}{(string): A vector of strings or symbols indicating the field names to use as "start" and "stop" temporal indices.
If only one value is given, that value will be repeated as the "stop" index}

\item{timeout}{The largest allowable 'gap' in a series of time values before a new 'island' begins: can be a quoted expression that conditionally determines the value.
If using date or datetime values for \code{time_fields}, specify the timeout using an appropriate \code{lubridate} functions (e.g., \code{\link[lubridate]{days}}).}

\item{boundary_name}{(string): The name root of the boundary column names (e.g., "episode" => "episode_start_idx", "episode_end_idx")}

\item{archipelago}{(logical | TRUE): Should the output include the islands and gaps generated?}

\item{show.all}{(logical | FALSE): Should the output include all of the columns of the output? \code{show.all} and \code{archipelago} are independent}
}
\value{
A \code{\link[data.table]{data.table}} with the following columns \code{<mapFields>, <X>, <Y>}:\cr
\describe{
\item{X (\code{archipelago = TRUE})}{\code{\{boundary_name\}_start_idx, \{boundary_name\}_end_idx, ISLAND, GAP}}
\item{Y (\code{show.all = TRUE})}{\code{island_idx, seq_idx, start_idx, stop_idx, rec_idx, map_partition, delta_start, delta_stop}}
}
}
\description{
\code{continuity} is based on the \href{https://www.red-gate.com/simple-talk/sql/t-sql-programming/the-sql-of-gaps-and-islands-in-sequences/}{'islands & gaps'} concept.
Use case for \code{continuity} is to roll up a sequence consisting of smaller duration into a larger epoch governed by some meaningful separation between \eqn{n_\text{lower}} and \eqn{{n - 1}_\text{upper}} segments.
}
\concept{Data Generation}
