% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_continuity.R
\name{continuity}
\alias{continuity}
\title{Continuity Creator}
\usage{
continuity(
  data,
  map_fields,
  time_fields,
  timeout = 0,
  boundary_name = "window",
  archipelago = TRUE,
  show.all = FALSE
)
}
\arguments{
\item{data}{(object): The source dataset, including all non-sessioning fields desired}

\item{map_fields}{(string): A comma-separated string literal containing field names that will partition \code{data}}

\item{time_fields}{(string): A comma-separated string literal containing field names to use as "start" and "stop" temporal indices.  If only one value is given, that value will be repeated as the "stop" index}

\item{timeout}{The largest allowable 'gap' in a series of time values before a new 'island' begins: can be a quoted expression that conditionally determines the value.}

\item{boundary_name}{(string): The name root of the boundary column names (e.g., "episode" -> "episode_start_idx", "episode_end_idx")}

\item{archipelago}{(logical | TRUE): Should the output include the islands and gaps generated?}

\item{show.all}{(logical | FALSE): Should the output include all of the columns of the output?}
}
\value{
A data.table with columns \if{html}{\out{<mapFields>}}, ..., timeOut, ISLAND, island_idx, where '...' is empty if \code{show.all} is FALSE
}
\description{
\code{continuity} is conceptually based on the \href{https://www.red-gate.com/simple-talk/sql/t-sql-programming/the-sql-of-gaps-and-islands-in-sequences/}{'islands & gaps'} concept.
}
\concept{Data Generation}
