% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_event.R
\name{Event}
\alias{Event}
\title{Create an Event}
\arguments{
\item{data}{(formula) A formula, where the LHS (if given) becomes the event label when argument \code{label} is not provided. The RHS becomes a reference to the source data by name. If a pipe (\code{`|`}) is given after the data source name on the RHS, it is treated as an indication of an expression that will filter rows in the data when evaluated.}

\item{jk, start, end(string)}{The name of the element in \code{data} denoting the \emph{'join-key'}, \emph{'start'}, and \emph{'end'} respectively.}

\item{label}{A unique label for the event source}
}
\value{
A list comprised of the event label and three quosures (\code{jk}, \code{time_start_idx}, \code{time_end_idx})
}
\description{
This funciton leverages the \code{lambda.r} library to create \code{Event} types comprised of a list of quosures to make referencing data sources and elements easy with \code{rlang::eval_tidy()}.
}
\examples{
\dontrun{
  library(event.vectors)
  Event(
    Event_B ~ evs_src_03
    , "join_key", "date.start", "date.end"
    )
  Event(
    "World Peace" ~ evs_src_01 | date.start >= "2026-02-15"
    , "join_key", "date.start", "date.end"
    )
}

}
