% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_documentation.R
\name{event.vectors-package}
\alias{event.vectors-package}
\title{Event Vectors}
\description{
Package \code{event.vectors} provides a framework for temporal data processing with a focus on capturing the relationships among points and durations in time.  Some definitions should be kept in mind to understand the event vector framework:
\itemize{
\item \code{Vector}: A vector numerically encodes magnitude and direction.
\item \code{Event}: Theoretically, an instantaneous point in time associated with a change of state.  Practically, an event is temporal duration encoded as a pair of temporal points. This is the understanding used in \code{event.vectors}.
}

The vector central to \code{event.vectors} is a complex vector that concisely describes various temporal relations between events.  The entire framework lends itself to temporal graph constructs in addition to tabular representations.  This allows for execution of queries and statistical analysis against the dataset.
}
