% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. otherFunctions.R
\name{make.evs_universe}
\alias{make.evs_universe}
\title{Create the EVSpace Universe}
\usage{
make.evs_universe(
  self,
  ...,
  time.control = list(-Inf, Inf),
  graph.control = NULL,
  omit.na = FALSE,
  chatty = FALSE
)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}) Logical expression that retain graph edges meeting the conditions}

\item{time.control}{A 2-element list containing the minimum and maximum values allowed for total temporal span between two events}

\item{graph.control}{An expression list containing \code{\link[igraph]{igraph-package}} calls to manipulate the internally-created graph in the order provided.  Use symbol \code{g} to generically denote the graph}

\item{omit.na}{(logical | \code{FALSE}) \code{TRUE} removes rows from class member \code{space} that have \code{NA} values}

\item{chatty}{(logical | \code{FALSE}) Verbosity flag}

\item{evs}{An \code{EVSpace} object}
}
\value{
Invisibly, the original object augmented with new member \code{space}
}
\description{
\code{make.evs_universe} supplies values to two class fields: \code{q_graph} and \code{space}, the latter being created from the former.
}
\section{Class Member \code{$space}}{

\code{$space} should have as many rows as the sum of all edge counts for graphs in \code{$q_graph}
}

\section{Parallelism}{

Parallelism over \code{$q_graph} is supported via package \href{https://rdocumentation.org/packages/furrr/versions/0.3.1}{furrr}: the user is responsible for setting a \code{\link[future]{plan}}.
}

